Backend - Added authentication handling for colection management

+++ src/main/java/io/dondany/fc/collection/CollectionService.java
@@ -2,12 +2,14 @@ package io.dondany.fc.collection;
 
 import io.dondany.fc.project.Project;
 import io.dondany.fc.project.ProjectRepository;
+import io.dondany.fc.user.User;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -15,21 +17,32 @@ public class CollectionService {
     private final CollectionRepository collectionRepository;
     private final ProjectRepository projectRepository;
 
-    List<CollectionDto> getAllCollectionsByProjectId(Long projectId) {
+    List<CollectionDto> getAllCollectionsByProjectId(Long projectId, User user) {
+        Project project = projectRepository.findById(projectId)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
+        if (!user.equals(project.getUser())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         return collectionRepository.findAllCollectionsByProjectId(projectId);
     }
 
 
-    public CollectionDto getOneById(Long projectId, Long id) {
+    public CollectionDto getOneById(Long projectId, Long id, User user) {
+        Project project = projectRepository.findById(projectId)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
+        if (!user.equals(project.getUser())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         return collectionRepository.findOne(projectId, id);
     }
 
-    public CollectionDto addCollection(Long projectId, Collection collection) {
-        Optional<Project> projectOptional = projectRepository.findById(projectId);
-        if (projectOptional.isEmpty()) {
-            throw new RuntimeException("Project with id " + projectId + " not found!");
+    public CollectionDto addCollection(Long projectId, Collection collection, User user) {
+        Project project = projectRepository.findById(projectId)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
+        if (!user.equals(project.getUser())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
         }
-        collection.setProject(projectOptional.get());
+        collection.setProject(project);
 
         Collection newCollection = collectionRepository.save(collection);
         return CollectionDto.builder()
@@ -41,26 +54,35 @@ public class CollectionService {
                 .build();
     }
 
-    public void deleteCollection(Long id) {
+    public void deleteCollection(Long projectId, Long id,  User user) {
+        Project project = projectRepository.findById(projectId)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
+        if (!user.equals(project.getUser())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         collectionRepository.deleteById(id);
     }
 
     @Transactional
-    public CollectionDto updateCollection(CollectionUpdateDto updateDto, Long projectId, Long id) {
-        Optional<Collection> existing = collectionRepository.findById(id);
-        if (existing.isEmpty()) {
-            throw new IllegalArgumentException("Can't find collection with id " + id);
+    public CollectionDto updateCollection(CollectionUpdateDto updateDto,
+                                          Long projectId,
+                                          Long id,
+                                          User user) {
+        Project project = projectRepository.findById(projectId)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
+        if (!user.equals(project.getUser())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
         }
 
-        Collection toUpdate = existing.get();
-        toUpdate.setName(updateDto.getName());
-        toUpdate.setDescription(updateDto.getDescription());
-
+        Collection collection = collectionRepository.findById(id)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
+        collection.setName(updateDto.getName());
+        collection.setDescription(updateDto.getDescription());
         return CollectionDto.builder()
-                .id(toUpdate.getId())
-                .name(toUpdate.getName())
-                .description(toUpdate.getDescription())
-                .project(toUpdate.getProject())
+                .id(collection.getId())
+                .name(collection.getName())
+                .description(collection.getDescription())
+                .project(collection.getProject())
                 .numberOfFlashCards(0L)
                 .build();
     }
